Moved session sagas to their own module.
